TIMOB-7878 Addressing issues mentioned.
