+++ support/iphone/run.py
@@ -1,25 +1,5 @@
 import os,subprocess,types,sys
-
-def ensure_dev_path():
-	rc = os.system("xcode-select -print-path >/dev/null 2>/dev/null")
-	if rc == 0 :
-		return
-	print '[INFO] XCode 4.3+ likely. Searching for developer folders.'
-	trypath = '/Developer'
-	if os.path.isdir(trypath):
-		os.putenv('DEVELOPER_DIR',trypath)
-		return
-	trypath = '/Applications/Xcode.app/Contents/Developer'
-	if os.path.isdir(trypath):
-		os.putenv('DEVELOPER_DIR',trypath)
-		return
-	spotlight_args = ['mdfind','kMDItemDisplayName==Xcode&&kMDItemKind==Application']
-	spotlight = subprocess.Popen(spotlight_args, stderr=subprocess.STDOUT, stdout=subprocess.PIPE)
-	for line in spotlight.stdout.readlines():
-		trypath = line.rstrip()+'/Contents/Developer'
-		if os.path.isdir(trypath):
-			os.putenv('DEVELOPER_DIR',trypath)
-			return
+from tools import ensure_dev_path
 
 def run(args,ignore_error=False,debug=True,out=None):
 	ensure_dev_path()
