+++ support/iphone/tools.py
@@ -6,6 +6,27 @@ template_dir = os.path.abspath(os.path.dirname(sys._getframe(0).f_code.co_filena
 
 from pbxproj import PBXProj
 
+def ensure_dev_path():
+	rc = os.system("xcode-select -print-path >/dev/null 2>/dev/null")
+	if rc == 0 :
+		return
+	print '[INFO] XCode 4.3+ likely. Searching for developer folders.'
+	trypath = '/Developer'
+	if os.path.isdir(trypath):
+		os.putenv('DEVELOPER_DIR',trypath)
+		return
+	trypath = '/Applications/Xcode.app/Contents/Developer'
+	if os.path.isdir(trypath):
+		os.putenv('DEVELOPER_DIR',trypath)
+		return
+	spotlight_args = ['mdfind','kMDItemDisplayName==Xcode&&kMDItemKind==Application']
+	spotlight = subprocess.Popen(spotlight_args, stderr=subprocess.STDOUT, stdout=subprocess.PIPE)
+	for line in spotlight.stdout.readlines():
+		trypath = line.rstrip()+'/Contents/Developer'
+		if os.path.isdir(trypath):
+			os.putenv('DEVELOPER_DIR',trypath)
+			return
+
 def read_config(f):
 	props = {}
 	if os.path.exists(f):
